apps/english_flashrace/requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
jinja2>=3.1
python-multipart>=0.0.9
python-dotenv>=1.0
itsdangerous>=2.1
